requirements.txt
Django==4.0.5
djangorestframework==3.13.1
Pillow==9.1.1
djangorestframework-simplejwt==5.1.0
ipdb==0.13.9
environ==1.0
psycopg2-binary==2.9.3
coverage==6.3.3
PyJWT==2.3.0
django-environ==0.9.0
requests==2.28.0
beautifulsoup4==4.11.1
lxml==4.9.0
selenium==4.3.0
django-elasticsearch==7.0.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
celery==5.2.7
redis==4.3.3
flower==1.0.0
django-celery-beat==2.3.0